refactor: 이벤트 삭제 및 의미없이 분리한 메서드 다시 합치기
