+++ src/main/java/com/woowa/woowakit/domain/order/domain/OrderRollbackService.java
@@ -3,6 +3,7 @@ package com.woowa.woowakit.domain.order.domain;
 import java.util.List;
 
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.woowa.woowakit.domain.cart.domain.CartItem;
@@ -11,18 +12,27 @@ import com.woowa.woowakit.domain.order.domain.mapper.CartItemMapper;
 import com.woowa.woowakit.domain.product.domain.product.Product;
 import com.woowa.woowakit.domain.product.domain.product.ProductRepository;
 
+import io.micrometer.core.annotation.Counted;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 
 @Component
 @RequiredArgsConstructor
+@Slf4j
 public class OrderRollbackService {
 
 	private final CartItemRepository cartItemRepository;
 	private final ProductRepository productRepository;
+	private final OrderRepository orderRepository;
 	private final CartItemMapper cartItemMapper;
 
-	@Transactional
-	public void rollback(final Order order) {
+	@Transactional(propagation = Propagation.REQUIRES_NEW)
+	@Counted("order.payment.failure")
+	public void rollback(final Long orderId, final Throwable error) {
+		log.error("결제 실패 복구 시작 orderId: {}, message={}", orderId, error.getMessage());
+
+		Order order = findOrderById(orderId);
+		order.cancel();
 		rollbackCartItems(order);
 		rollbackProducts(order);
 	}
@@ -43,5 +53,9 @@ public class OrderRollbackService {
 	private void rollbackCartItems(final Order order) {
 		List<CartItem> cartItems = cartItemMapper.mapAllFrom(order);
 		cartItemRepository.saveAll(cartItems);
+	}
+
+	private Order findOrderById(final Long id) {
+		return orderRepository.findById(id).orElseThrow();
 	}
 }
